movie-recommandation-turkish/requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langgraph
chromadb
transformers
sentence-transformers
torch
pydantic
tqdm
openai